% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaBetaContInPlasma.R
\name{alphaBetaContInPlasma}
\alias{alphaBetaContInPlasma}
\title{alphaBetaContInPlasma}
\usage{
alphaBetaContInPlasma(psPlByBlock, psallzeroInNC, blk, alphaBetaNegControl)
}
\arguments{
\item{psPlByBlock}{list, phyloseq objects with plasma samples.}

\item{psallzeroInNC}{list, phyloseq objects with all taxa with no prevalence in negative controls.}

\item{blk}{numeric, which block samples to be used.}

\item{alphaBetaNegControl}{list, estimated contamination distribution parameters using all negative control samples.}
}
\value{
list, estimated contamination distribution parameters for the plasma samples.
}
\description{
alphaBetaContInPlasma
}
