% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BARBI.R
\docType{package}
\name{BARBI}
\alias{BARBI}
\title{BARBI: BAyesian Reduction in Background Interference}
\description{
A package for distinguishing true taxa from DNA contamination in molecular microbial studies.
The BARBI package provides three important functions: alphaBetaNegControl, alphaBetaContInPlasma and samplingPosterior.
}
\section{alphaBetaNegControl function}{

        The alphaBetaNegControl function estimates the parameters of the distribution of the contamination intensity in negtiave control samples.
}

\section{alphaBetaContInPlasma function}{

        The alphaBetaContInPlasma function estimates the parameters of the distribution of the contamination intensity in specimen samples.
}

\section{samplingPosterior fucntion}{

        The samplingPosterior function samples from the posterior distribution of the true intensity.
}

