% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAtopicmodel.R
\name{LDAtopicmodel}
\alias{LDAtopicmodel}
\title{Bayesian latent Dirichlet allocation for 16S and metagenomics data.}
\usage{
LDAtopicmodel(stan_data, ...)
}
\arguments{
\item{stan_data}{A list of stan data.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian latent Dirichlet allocation for 16S and metagenomics data.
}
