% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTable.R
\name{makeTable}
\alias{makeTable}
\title{Make summaires from the BARBI results}
\usage{
makeTable(itera, burnIn, num_blks, cov.pro, psByBlock, gammaPrior_all_blks,
  post_all_blocks, mak_tab = TRUE)
}
\arguments{
\item{itera}{numeric, number of MCMC samples.}

\item{burnIn}{numeric, number of states from MCMC to discard.}

\item{num_blks}{numeric, number of blocks.}

\item{cov.pro}{numeric, coverage probability to construct the highest posterior density interval.}

\item{psByBlock}{list, phyloseq objects by blocks.}

\item{gammaPrior_all_blks}{list, estimated distribution parameters for the intensity of contamination in the plasma samples in all blocks.}

\item{post_all_blocks}{list, posterior results}

\item{mak_tab}{logical, whether make a table for each sample with true signal.}
}
\value{
tables with true signal and a list of true signal
}
\description{
Make summaires from the BARBI results
}
