% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_MCMC.R
\name{MH_MCMC}
\alias{MH_MCMC}
\title{MH_MCMC}
\usage{
MH_MCMC(iterations, k, al_c, be_c, startvalue_lamda_r = 0)
}
\arguments{
\item{iterations}{numeric, number of MCMC}

\item{k}{numeric vector, observed reads for all taxa in a sample}

\item{al_c}{numeric vector, estimated shape parameter for the contamination distribution}

\item{be_c}{numeric vector, estimated scale parameter for the contamination distribution}
}
\value{
array
}
\description{
MH_MCMC
}
