% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposalFunction.R
\name{proposalFunction}
\alias{proposalFunction}
\title{proposalFunction}
\usage{
proposalFunction(al_p, be_p)
}
\arguments{
\item{al_p}{numeric, shape parameter for the proposed density function}

\item{be_p}{numeric, scale parameter for the proposed density function}
}
\value{
numeric, proposed density function
}
\description{
Proposed gamma density function
}
