% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psBlockResults.R
\name{psBlockResults}
\alias{psBlockResults}
\title{psBlockResults}
\usage{
psBlockResults(ps, sampleTypeVar = "Sample_Type", blockVar = "block")
}
\arguments{
\item{ps}{phyloseq object}

\item{sampleTypeVar}{factor, with two levels "Plasma" and "Control" samples}

\item{blockVar}{factor, block of the samples}
}
\value{
list, a list of phyloseq by block, phyloseq of negative control samples by block, phyloseq of taxa with prevalence zero in negative control samples by block, phyloseq of plasma samples by block
}
\description{
psBlockResults
}
