% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psSub.R
\docType{data}
\name{psSub}
\alias{psSub}
\title{Simulated \code{"phyloseq"} object.}
\format{An object of class \code{"phyloseq"}.
\describe{
  \item{SubjectID}{IDs of individuals}
  \item{SampleType}{A factor with three levels: SIRS, Control, and Healthy}
  \item{Sample_Type}{A factor with two levels: Plasma, Control}
  \item{SampleCode}{Sample ID for sample}
  }}
\source{
A subset of samples from \href{https://academic.oup.com/ofid/article/4/suppl_1/S71/4294264}{where}
}
\usage{
psSub
}
\description{
A \code{"phyloseq"} class object with 1644 taxa from 268 SIRS, healthy, and negative control samples.
}
\examples{
\dontrun{sample_data(psSub)
otu_table(psSub)}
}
\keyword{datasets}
