% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingPosterior.R
\name{samplingPosterior}
\alias{samplingPosterior}
\title{samplingPosterior}
\usage{
samplingPosterior(psPlByBlock, blk, gammPrior, iter = 10000)
}
\arguments{
\item{psPlByBlock}{list, phyloseq objects with plasma samples.}

\item{blk}{numeric, which block samples to be used}

\item{gammPrior}{list, contamination distribution in the plasma samples}

\item{iter}{numeric, number of MCMC}
}
\value{
list, estiamted reference posterior for real reads in all samples
}
\description{
samplingPosterior
}
