% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaBetaContInPlasma.R
\name{alphaBetaContInPlasma}
\alias{alphaBetaContInPlasma}
\title{Compute gamma distribution parameters in plasma samples}
\usage{
alphaBetaContInPlasma(psPlByBlock, psallzeroInNC, blk, alphaBetaNegControl,
  stringent = FALSE)
}
\arguments{
\item{psPlByBlock}{A list of phyloseq objects with plasma samples.}

\item{psallzeroInNC}{A list of phyloseq objects with all taxa with no
prevalence in negative control samples.}

\item{blk}{Numeric, which block samples to be used.}

\item{alphaBetaNegControl}{A list of estimated distribution parameters for the intensity of contamination in negative control samples returned by \code{alphaBetaNegControl}.}

\item{stringent}{Logical, TRUE/FALSE: (i) TRUE keep the gamma parametersas in the negative controls for the taxa with mean less than 1, and (ii) FALSE use the scaling property of the gamma distribution.}
}
\value{
A list of estimated distribution parameters for the intensity of contamination in the plasma samples in a selected block.
}
\description{
Given the estimated gamma distribution parameters for the intensity of contamination in negative control samples, \code{alphaBetaContInPlasma}
computes the
gamma distribution parameters in the plasma sample using the scaling
property of the gamma distribution.
}
\section{Note}{

If all the samples are from one block, create a variable with one block
level.
}

\seealso{
\code{\link{alphaBetaNegControl}}
}
