% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaBetaNegControl.R
\name{alphaBetaNegControl}
\alias{alphaBetaNegControl}
\title{Compute gamma distirbution parameters}
\usage{
alphaBetaNegControl(psNCbyBlock)
}
\arguments{
\item{psNCbyBlock}{A list of phyloseq objects with all negative control
samples in each block.}
}
\value{
A list of estimated distribution parameters of contamination
        intensity in negative control samples.
}
\description{
\code{alphaBetaNegControl} computes the gamma distribution parameters
     using negative control samples for all blocks.
}
\examples{
\dontrun{
alphaBetaNegControl(psNCByBlock)
}
}
\seealso{
\code{\link[phyloseq]{phyloseq}}
}
