% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaBetaNegControl.R
\name{alphaBetaNegControl}
\alias{alphaBetaNegControl}
\title{alphaBetaNegControl}
\usage{
alphaBetaNegControl(psNCbyBlock)
}
\arguments{
\item{psNCbyBlock}{list, phyloseq objects with all negative control samples in each block.}
}
\value{
list, estimated contamination distribution parameters using all negative control samples.
}
\description{
alphaBetaNegControl
}
