% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaBetaNegControl.R
\name{alphaBetaNegControl}
\alias{alphaBetaNegControl}
\title{Compute gamma distirbution parameters}
\usage{
alphaBetaNegControl(psNCbyBlock, add1 = TRUE)
}
\arguments{
\item{psNCbyBlock}{A list of phyloseq objects with all negative control
samples in each block.}

\item{add1}{logical, whether adding one to the otu table in the phyloseq}
}
\value{
A list of estimated distribution parameters for the intensity of contamination in negative control samples.
}
\description{
Computes the gamma distribution parameters for the intensity of contamination using negative control samples in each block.
}
\examples{
\dontrun{
alphaBetaNegControl(psNCByBlock)
}
}
