% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_MCMC.R
\name{MH_MCMC}
\alias{MH_MCMC}
\title{MH_MCMC}
\usage{
MH_MCMC(iterations, k, al, be)
}
\arguments{
\item{iterations}{numeric, number of MCMC}

\item{k}{numeric vector, observed reads for all taxa in a sample}

\item{al}{numeric vector, estimated shape parameter for the contamination distribution}

\item{be}{numeric vector, estimated scale parameter for the contamination distribution}
}
\value{
array
}
\description{
MH_MCMC
}
