% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior}
\alias{posterior}
\title{posterior}
\usage{
posterior(lambda_r, k, al_c, be_c)
}
\arguments{
\item{lambda_r}{numeric, intensity of true signal.}

\item{k}{numeric, observed count for a taxa in a sample.}

\item{al_c}{numeric, estimated shape parameter for the intensity of contamination for a taxa.}

\item{be_c}{numeric, estimated scale parameter for the intensity of contamination for a taxa.}
}
\value{
numeric, posterior density function
}
\description{
Posterior density function
}
