% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior}
\alias{posterior}
\title{posterior}
\usage{
posterior(lambda_r, k, al_c, be_c)
}
\arguments{
\item{lambda_r}{numeric, intensity of real read.}

\item{k}{numeric, observed read for a taxa in a sample.}

\item{al_c}{numeric, estimated shape parameter for the contamination distribution for each taxa.}

\item{be_c}{numeric, estimated scale parameter for the contamination distribution for each taxa.}
}
\value{
numeric, posterior density
}
\description{
posterior
}
