% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior}
\alias{posterior}
\title{posterior}
\usage{
posterior(s, k, al, be)
}
\arguments{
\item{s}{numeric, intensity of real read.}

\item{k}{numeric, observed read for a taxa in a sample.}

\item{al}{numeric, estimated shape parameter for the contamination distribution for each taxa.}

\item{be}{numeric, estimated scale parameter for the contamination distribution for each taxa.}
}
\value{
numeric, posterior density
}
\description{
posterior
}
