% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposalFunction.R
\name{proposalFunction}
\alias{proposalFunction}
\title{proposalFunction}
\usage{
proposalFunction(al_p, be_p)
}
\arguments{
\item{param}{numeric, a proposal value}

\item{al.p}{numeric, shape parameter of the proposal function}

\item{be.p}{numeric, scale parameter of the proposal function}
}
\value{
numeric, proposal density
}
\description{
proposalFunction
}
