% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psBlockResults.R
\name{psBlockResults}
\alias{psBlockResults}
\title{psBlockResults}
\usage{
psBlockResults(ps, sampleTypeVar = "Sample_Type",
  sampleName = "SampleCode", blockVar = "block")
}
\arguments{
\item{ps}{phyloseq object.}

\item{sampleTypeVar}{character, variable defines the "Plasma" and "Control" samples.}

\item{sampleName}{character, variable defines the sample names.}

\item{blockVar}{character, variable defines the block of the samples. If there is only one block, create a variable with only one level.}
}
\value{
A list of phyloseq objects, phyloseq of negative control samples by block, phyloseq of taxa with prevalence zero in negative control samples by block, phyloseq of plasma samples by block.
}
\description{
Prepare the phyloseq object for the Bayesian inference
}
