% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingPosterior.R
\name{samplingPosterior}
\alias{samplingPosterior}
\title{Sampling posterior}
\usage{
samplingPosterior(psPlByBlock, blk, gammaPrior_Cont, itera = 10000)
}
\arguments{
\item{psPlByBlock}{A list of phyloseq objects with plasma samples.}

\item{blk}{Numeric, which block samples to be used.}

\item{gammaPrior_Cont}{A list of estimated distribution parameters for the
contamination intensity in thespecimen samples in a particular
block.}

\item{itera}{Numeric, number of MCMC samples.}
}
\value{
A list of marginal posterior for the true intensity
}
\description{
Sampling from the marginal posterior for the true intensity
}
