% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingPosterior.R
\name{samplingPosterior}
\alias{samplingPosterior}
\title{Sampling posterior}
\usage{
samplingPosterior(psPlByBlock, blk, gammaPrior_Cont, itera = 10000,
  ncores = ncores)
}
\arguments{
\item{psPlByBlock}{A list of phyloseq objects with plasma samples.}

\item{blk}{Numeric, which block samples to be used.}

\item{gammaPrior_Cont}{A list of estimated distribution parameters of
contamination intensity in the plasma samples in a particular
block.}

\item{itera}{Numeric, number of MCMC samples.}
}
\value{
A list of estimated marginal reference posterior for the intensity of true signal.
}
\description{
Sampling from the marginal reference posterior for the intensity of true signal
}
